+++ pkg/providers/mtlstruststore/testdata/TestDataCASetAssociations/association_type_enrollments.tf
@@ -1,0 +1,8 @@
+provider "akamai" {
+  edgerc = "../../common/testutils/edgerc"
+}
+
+data "akamai_mtlstruststore_ca_set_associations" "test" {
+  id               = "123"
+  association_type = "enrollments"
+}